14_Authentications/OAuth2_OR_OpenID_Connect/app/auth/config.py
import os
from datetime import timedelta
from dotenv import load_dotenv

load_dotenv()

# JWT
JWT_SECRET = os.getenv("JWT_SECRET", "CHANGE_ME")
JWT_ALGORITHM = os.getenv("JWT_ALGORITHM", "HS256")
ACCESS_EXPIRES = timedelta(minutes=int(os.getenv("ACCESS_MINUTES", "15")))
REFRESH_EXPIRES = timedelta(days=int(os.getenv("REFRESH_DAYS", "7")))

# Refresh cookie
REFRESH_COOKIE_NAME = "refresh_token"
REFRESH_COOKIE_PATH = "/auth"
REFRESH_COOKIE_HTTPONLY = True
REFRESH_COOKIE_SAMESITE = "strict"
REFRESH_COOKIE_SECURE = os.getenv("REFRESH_COOKIE_SECURE", "false").lower() == "true"

# Google OAuth
GOOGLE_CLIENT_ID = os.getenv("GOOGLE_CLIENT_ID", "")
GOOGLE_CLIENT_SECRET = os.getenv("GOOGLE_CLIENT_SECRET", "")
GOOGLE_REDIRECT_URI = os.getenv("GOOGLE_REDIRECT_URI", "http://127.0.0.1:8000/auth/callback/google")
